# public/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "QuickDCP API",
    "version": "0.1.0",
    "description": "Private API for QuickDCP System Core: jobs, upload, proof chain, verification, billing and worker control."
  },
  "servers": [
    {
      "url": "https://api.quickdcp.local",
      "description": "Local QuickDCP node"
    }
  ],
  "components": {
    "securitySchemes": {
      "Customer": {
        "type": "apiKey",
        "in": "header",
        "name": "X-QD-Customer",
        "description": "Customer identifier for multi-tenant setups."
      },
      "ApiKeyAuth": {
        "type": "apiKey",
        "in": "header",
        "name": "Authorization",
        "description": "QuickDCP API key, formatted as `QuickDCP <api-key>`."
      }
    },
    "schemas": {
...
      "ManifestQC": {
        "type": "object",
        "properties": {
          "audio_lufs": {
            "type": "number",
            "nullable": true
          },
          "video_issues": {
            "type": "integer",
            "nullable": true
          },
          "subtitle_sync_ms": {
            "type": "integer",
            "nullable": true
          }
        }
      },
      "Manifest": {
        "type": "object",
        "properties": {
          "job_id": {
            "type": "string"
          },
          "profile": {
            "type": "object",
            "additionalProperties": true
          },
          "outputs": {
            "type": "object",
            "additionalProperties": true
          },
          "qc": {
            "$ref": "#/components/schemas/ManifestQC"
          },
          "proof": {
            "type": "object",
            "additionalProperties": true
          }
        },
        "required": [
          "job_id"
        ]
      },
      "WorkerNextResponse": {
        "type": "object",
        "properties": {
          "job_id": {
            "type": "string",
            "nullable": true
          },
          "profile": {
            "type": "object",
            "nullable": true,
            "additionalProperties": true
          },
          "status": {
            "type": "string",
            "description": "EMPTY, QUEUED, PROCESSING"
          }
        },
        "required": [
          "status"
        ]
      },
      "WorkerUpdateRequest": {
        "type": "object",
        "properties": {
          "job_id": {
            "type": "string"
          },
          "manifest": {
            "$ref": "#/components/schemas/Manifest"
          },
          "status": {
            "type": "string",
            "description": "PASS, FAIL, PROCESSING, QUEUED"
          }
        },
        "required": [
          "job_id",
          "manifest",
          "status"
        ]
      },
...
    }
  },
  "security": [
    {
      "Customer": [],
      "ApiKeyAuth": []
    }
  ],
  "paths": {
...
    "/internal/next-job": {
      "post": {
        "summary": "Worker: fetch next job",
        "operationId": "workerNextJob",
        "parameters": [
          {
            "name": "X-Worker-Token",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Next job for worker, or EMPTY.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WorkerNextResponse"
                }
              }
            }
          },
          "401": {
            "description": "Bad token"
          }
        },
        "security": []
      }
    },
    "/internal/update-job": {
      "post": {
        "summary": "Worker: update job manifest and status",
        "operationId": "workerUpdateJob",
        "parameters": [
          {
            "name": "X-Worker-Token",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/WorkerUpdateRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Job updated"
          },
          "401": {
            "description": "Bad token"
          },
          "404": {
            "description": "Job not found"
          }
        },
        "security": []
      }
    },
...
  }
}